requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
redis
mysql-connector-python